+++ bin/build-shaders.sh
@@ -20,7 +20,7 @@ SHADER_DIRS=(
     "src/contents/ui/effects/rainbow-waves/shaders"
     "src/contents/ui/effects/lava-lamp/shaders"
     "src/contents/ui/effects/dot-waves/shaders"
-    "src/contents/ui/effects/trippy-tunnel/shaders"
+    "src/contents/ui/effects/vortex/shaders"
     "src/contents/ui/filters/shaders"
 )